+++ src/qcommon/vm_armv7l.c
@@ -3219,7 +3219,7 @@ __recompile:
 
 int32_t VM_CallCompiled( vm_t *vm, qint nargs, int32_t *args )
 {
-	int32_t		opStack[ MAX_OPSTACK_SIZE ];
+	static int32_t		opStack[ MAX_OPSTACK_SIZE ];
 	int32_t		stackOnEntry;
 	int32_t		*image;
 	qint		i;